fix(windows): correct type mismatches for windows 0.58 API
- Cast WorkingSetSize to u64 in get_process_memory
- Use PWSTR::null() instead of PCWSTR::null() for failure_actions

+++ crates/agent-core/src/resources.rs
@@ -507,7 +507,7 @@ fn get_process_memory() -> u64 {
         pmc.cb = std::mem::size_of::<PROCESS_MEMORY_COUNTERS>() as u32;
 
         if GetProcessMemoryInfo(process, &mut pmc, pmc.cb).is_ok() {
-            return pmc.WorkingSetSize;
+            return pmc.WorkingSetSize as u64;
         }
     }
     0
